refactor: move list items to components folder
