+++ src/app/Scenes/Onboarding/OnboardingPersonalization2/Components/PartnerListItem.tsx
